Rename atomic module to transaction module
